Use core-remoteviews library to simplify code
- Use the RemoteViewsCompat adapter
- Use the new size mode
- Clean samples
