+++ AppWidget/app/src/main/java/com/example/android/appwidget/list/ListAppWidget.kt
@@ -22,9 +22,10 @@ import android.appwidget.AppWidgetManager
 import android.appwidget.AppWidgetProvider
 import android.content.Context
 import android.content.Intent
-import android.util.SizeF
 import android.widget.RemoteViews
 import androidx.annotation.LayoutRes
+import androidx.core.util.SizeFCompat
+import androidx.core.widget.createResponsiveSizeAppWidget
 import com.example.android.appwidget.MainActivity
 import com.example.android.appwidget.R
 
@@ -97,18 +98,17 @@ class ListAppWidget : AppWidgetProvider() {
             val remoteViews = if (layoutId == R.layout.widget_grocery_list) {
                 // Specify the maximum width and height in dp and a layout, which you want to use
                 // for the specified size
-                val viewMapping = mapOf(
-                    SizeF(150f, 150f) to constructRemoteViews(
+                val sizes = listOf(SizeFCompat(150f, 150f), SizeFCompat(250f, 150f))
+                createResponsiveSizeAppWidget(appWidgetManager, appWidgetId, sizes) { size ->
+                    val id = if (size == sizes[0]) {
                         R.layout.widget_grocery_list
-                    ), SizeF(250f, 150f) to constructRemoteViews(
+                    } else {
                         R.layout.widget_grocery_grid
-                    )
-                )
-                RemoteViews(viewMapping)
+                    }
+                    constructRemoteViews(id)
+                }
             } else {
-                constructRemoteViews(
-                    layoutId
-                )
+                constructRemoteViews(layoutId)
             }
             appWidgetManager.updateAppWidget(appWidgetId, remoteViews)
         }
