+++ AppWidget/app/src/main/res/drawable-nodpi/buttons_widget_preview.png
