+++ AppWidget/app/src/main/res/xml/app_widget_buttons.xml
@@ -20,13 +20,13 @@ reflect the inflated contents of ListView in a widget preview.
 Thus, only supplying a static image through android:previewImage
 -->
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-    android:description="@string/items_widget_description"
-    android:initialKeyguardLayout="@layout/widget_items_collection"
-    android:initialLayout="@layout/widget_items_collection"
+    android:description="@string/buttons_widget_description"
+    android:initialKeyguardLayout="@layout/widget_buttons"
+    android:initialLayout="@layout/widget_buttons"
     android:minWidth="180dp"
     android:minHeight="220dp"
-    android:previewImage="@drawable/items_collection_widget_preview"
-    android:previewLayout="@layout/widget_items_collection_preview"
+    android:previewImage="@drawable/buttons_widget_preview"
+    android:previewLayout="@layout/widget_buttons_preview"
     android:resizeMode="horizontal|vertical"
     android:targetCellWidth="3"
     android:targetCellHeight="4"
